Tarea 1_A01740797

+++ assignments/02AreaTriangulo/src/exercise.py
@@ -1,7 +1,7 @@
 def main():
     #escribe tu código abajo de esta línea
-    n1 = int(input('Dame la base: '))
-    n2 = int(input('Dame la altura: '))
+    n1 = float(input('Dame la base: '))
+    n2 = float(input('Dame la altura: '))
 
     área = ( n1 * n2 )/ 2
