+++ assignments/03Promedio/src/exercise.py
@@ -1,9 +1,9 @@
 def main():
     #escribe tu código abajo de esta línea
-    n1 = int(input('Calificación de la materia: '))
-    n2 = int(input('Calificación de la materia: '))
-    n3 = int(input('Calificación de la materia: '))
-    n4 = int(input('Calificación de la materia: '))
+    n1 = float(input('Calificación de la materia: '))
+    n2 = float(input('Calificación de la materia: '))
+    n3 = float(input('Calificación de la materia: '))
+    n4 = float(input('Calificación de la materia: '))
 
     promedio = ( n1 + n2 + n3 + n4 ) / 4
